Adds Basic JSON REST Interface
Adds Maven dependencies for a basic JSON Restful interface and adds a
new class which acts as the service. New API routes are added to 
/api/v1 which serves both an auth endpoint and a tasks endpoint. Basic
routes to find all, creation of a new task and display a particular
task were added. TODO add delete and update routes and move database
connection logic to separate class.